+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/SolicitorCreateContestedMidHandlerTest.java
@@ -6,8 +6,8 @@ import org.junit.Test;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.finrem.caseorchestration.ccd.callback.CallbackType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.controllers.GenericAboutToStartOrSubmitCallbackResponse;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.CaseType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.ContestedChildrenDetailDataWrapper;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.ContestedChildrenService;
